// eLog.Web/wwwroot/src/app/Models/company/company.module.ts
export interface Company {
    companyID: number;
    companyName: string;
    companyAbbreviation: string;
    phoneNumber: string;
    extension: string;
    faxNumber: string;
    emailAddress: string;
    website: string;
    country: string;
    countryCode: string;
    state: string;
    city: string;
    address: string;
    imageName: string;
    imageFilePath: string;
    createdBy: number;
    createdTS: string;
    modifiedBy: number;
    
}